+++ lib/x_label_config.dart
@@ -2,13 +2,13 @@ part of stacked_bar_chart;
 
 class XLabelConfiguration {
   ///Add styling to the X-axis labels
-  final TextStyle xLabelStyle;
+  final TextStyle labelStyle;
 
   ///This parameter maps the values as X-Axis labels
-  final String Function(DateTime p1) xLabelMapper;
+  final String Function(DateTime p1) labelMapper;
 
   XLabelConfiguration({
-    this.xLabelStyle,
-    this.xLabelMapper,
+    this.labelStyle,
+    this.labelMapper,
   });
 }